# Conflicts:
#	src/libraries/Libraries.Services/DevQuiz.Libraries.Services.xml
#	src/libraries/Libraries.Services/UserService.cs

+++ src/TelegramBot/Mappers/UserBotMapperProfile.cs
@@ -24,7 +24,7 @@ namespace DevQuiz.TelegramBot.Mappers
 
             CreateMap<Chat, TUserDto>()
                 .ForMember(dest => dest.Id, opt => opt.Ignore())
-                .ForMember(dest => dest.TelegramChatId, opt => opt.MapFrom(src => (int)src.Id))
+                .ForMember(dest => dest.TelegramChatId, opt => opt.MapFrom(src => src.Id))
                 .ForMember(dest => dest.UserName, opt => opt.MapFrom(src => src.Username))
                 .ForMember(dest => dest.FirstName, opt => opt.MapFrom(src => src.FirstName))
                 .ForMember(dest => dest.LastName, opt => opt.MapFrom(src => src.LastName))
