+++ src/TelegramBot/MediatR/Handlers/StartCommandHandler.cs
@@ -61,7 +61,7 @@ namespace DevQuiz.TelegramBot.MediatR.Handlers
             _logger.LogInformation($"{nameof(StartCommand)}Handle begins for chatId={_chat.Id}.");
             _cancellationToken = cancellationToken;
 
-            var userInDb =  await _userService.GetByChatIdAsync((int)_chat.Id, cancellationToken);
+            var userInDb =  await _userService.GetByChatIdAsync(_chat.Id, cancellationToken);
 
             userInDb = await CheckUserAsync(userInDb);
